refactor: standardize paths in require()'s
