+++ controllers/errorsController.js
@@ -2,8 +2,8 @@ const path = require('path')
 const paths = require('./pathsController')
 const ClientError = require('./utils/ClientError')
 const ServerError = require('./utils/ServerError')
-const config = require('../config')
-const logger = require('../logger')
+const config = require('./../config')
+const logger = require('./../logger')
 
 const self = {
   errorPagesCodes: Object.keys(config.errorPages)
