+++ routes/file.js
@@ -1,7 +1,7 @@
 const { Router } = require('hyper-express')
 const routes = new Router()
-const utils = require('../controllers/utilsController')
-const config = require('../config')
+const utils = require('./../controllers/utilsController')
+const config = require('./../config')
 
 routes.get('/file/:identifier', async (req, res) => {
   // Uploads identifiers parsing, etc., are strictly handled by client-side JS at src/js/file.js
